{"schema":"decodex/commit/1","summary":"Consolidate cohesive Rust type groups","authority":"manual"}
